Fix module naming conflicts and server initialization
